Update BSP library and sample code for Keil AC6
1. Add *.uvprojx to the Keil folder of each example.
2. Fix the build problems on Keil V5.28 and V5.38.
3. Fix the build problems on IAR V8.40.
4. Fix the build problems on GCC.

commit again

+++ SampleCode/StdDriver/I2C_Wakeup_Slave/main.c
@@ -313,7 +313,7 @@ int32_t main(void)
     while((g_u8SlvPWRDNWK & g_u8SlvI2CWK) == 0);
 
     /* Wake-up Interrupt Message */
-    printf("Power-down Wake-up INT 0x%x\n", ((CLK->PWRCTL) & CLK_PWRCTL_PDWKIF_Msk));
+		printf("Power-down Wake-up INT 0x%x\n", (uint32_t)(((CLK->PWRCTL) & CLK_PWRCTL_PDWKIF_Msk)));
     printf("I2C0 WAKE INT 0x%x\n", I2C0->WKSTS);
 
     /* Disable power wake-up interrupt */
